+++ test/test_users.rb
@@ -28,6 +28,14 @@ class UserTestCase < TestCaseWithSplunkConnection
   def test_case_insensitive
     name = temporary_name() + "UPCASE"
     user = @service.users.create(name, :password => "abc", :roles => ["power"])
+    assert_equal(name.downcase(), @service.users.fetch(name).name)
     assert_true(@service.users.has_key?(name.downcase()))
+  end
+
+  def test_roles
+    name = temporary_name()
+    user = @service.users.create(name, :password => "abc",
+                                 :roles => ["power", "can_delete"])
+    assert_equal(["can_delete", "power"], user["roles"].sort())
   end
 end